moving towards composer, WIP, issue #14
